Stats: Moving UTM files to new directory
